Several bug fixes and code cleanup
Reuse existing connection in the brokers when connecting to the parents or children instead of creating a new one every time.

Fix bug in which a broker only sends a publication to the first matching subscriber instead of all.

Fix bug in which the publication's originURL was not changed when sending in Filter mode.